test(e2e/tron): refactor network tests for speed
